Move state to Routes component
